+++ src/app/layout.tsx
@@ -13,6 +13,7 @@ import './plugin';
 import NaverAnalytics from '@/lib/NaverAnalytics';
 import AuthWrapper from '@/components/common/AuthWrapper';
 import ToastProvider from '@/providers/ToastProvider';
+import RefreshTokenWrapper from '@/components/common/RefreshTokenWrapper';
 
 declare global {
   interface Window {
@@ -100,6 +101,7 @@ export default function RootLayout({
             />
             {/* <AuthWrapper /> */}
             <GnbWrapper />
+            <RefreshTokenWrapper />
             {children}
             <MobileFooter />
             <ModalProvider />
