+++ src/components/product/detail/Notice.tsx
@@ -1,6 +1,6 @@
 import NoticeItemSkeleton from '@/components/skeleton/NoticeItemSkeleton';
 import UseNoticeLists from '@/factory/useNoticeLists';
-import { INoticeItem } from '@/types/BuildingProductType';
+import { INoticeItem } from '@/types/Product/BuildingProductType';
 import { useCallback, useMemo } from 'react';
 import { Virtuoso } from 'react-virtuoso';
