+++ src/components/product/detail/Report.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image';
 import { formatDate } from '@/utils/FormatDate';
 import UseReportList from '@/factory/useReportList';
 import { useCallback } from 'react';
-import { IreportItem } from '@/types/BuildingProductType';
+import { IreportItem } from '@/types/Product/BuildingProductType';
 import NoticeItemSkeleton from '@/components/skeleton/NoticeItemSkeleton';
 import { Virtuoso } from 'react-virtuoso';
 const Report = () => {
