+++ src/components/product/detail/music/chart/MusicStockPriceChart.tsx
@@ -3,7 +3,7 @@ import { Line } from 'react-chartjs-2';
 
 import { useQuery } from '@tanstack/react-query';
 import { usePathname } from 'next/navigation';
-import { IMusicBulidingStockPriceChart } from '@/types/MusicProductType';
+import { IMusicBulidingStockPriceChart } from '@/types/Product/MusicProductType';
 import { basicAxiosInstance } from '@/service/axiosInstance';
 import { TooltipItem } from 'chart.js';
