+++ src/components/product/detail/music/chart/YoutubeViewChart.tsx
@@ -2,7 +2,7 @@ import { useQuery } from '@tanstack/react-query';
 import { usePathname } from 'next/navigation';
 import { Line } from 'react-chartjs-2';
 import { useRef, useState } from 'react';
-import { IContentYoutubeViewCharts } from '@/types/MusicProductType';
+import { IContentYoutubeViewCharts } from '@/types/Product/MusicProductType';
 import { basicAxiosInstance } from '@/service/axiosInstance';
 import { TooltipItem } from 'chart.js';
