+++ src/components/quiz/QuizBanner.tsx
