+++ src/factory/Product/ProductDetail/BuildingProductDetail.ts
@@ -1,5 +1,5 @@
 import { axiosInstance } from '@/service/axiosInstance';
-import { IBuildingProductDetail } from '@/types/BuildingProductType';
+import { IBuildingProductDetail } from '@/types/Product/BuildingProductType';
 import { useQuery, QueryFunctionContext } from '@tanstack/react-query';
 
 const fetchBuildingProductDetail = async (context: QueryFunctionContext) => {
