+++ src/types/Product/ContentProductType.ts
