+++ src/types/Product/CowProductType.ts
